Test too few arguments in CLI.add and fix found issue. Update PyDoc.
CLI.add already had supposed functionality to recognize and catch cases
of too few arguments given. Unfortunately, this functionality was
untested, flawed and thus ineffective.  It has been fixed and updated
with an according test.

Additionally, the PyDoc of the method has been updated to account for
the fact that it does not make sense to provide a gap-handling option in
this method.