Fix bug in FileSet.add_file_set with option strip_gaps
The method contained a bug that caused gaps that were wider than just
one file to not be stripped completely. In theory, it could have also
prevented actual files from being added. The bug was caused by the
removal of elements from an index list while a for loop was still
iterating over that list at the same time. It was fixed by introducing a
separate list to add the actual indexes to be added to during runtime.
An additional add_file_set unit test was created to catch such faulty
behavior in the future.