Fix issue in CLI.rename with multiple file sets existing
If multiple known file sets exist in the file_set_cache, each of them is
checked against the given new_pattern. Unfortunately, due to bad
variable naming, the for-loop overrode the file_set parameter, thus
performing the operation on whichever file set was found last in the
file set cache. The variable has been renamed to avoid this problem.