Fix issue in FileSet.add_file_set with multi-assigned indexes
The method add_file_set had issues handling multi-assigned indexes.
Whenever one was encountered, there was an iteration over the list of
file types at this index. Unfortunately, since this list was being
updated already within the for loop, it lead to side effects such as
files being skipped.

This issue has been solved using a shallow copy of the list and a unit
test has been added to catch such an issue in the future.