Fix change_pattern tests for FileSet
Previously, the "test_rename.py" module was error ridden, as the FileSet
had since moved on to call the method under test "change_pattern". Since
the test module wasn't run by the runner due to a name conflict with
CLI_tests.test_rename, this error was never resolved. Until now, that
is.